consolidate files under better names
